+++ markets/bfp-market/contracts/interfaces/IOrderModule.sol
@@ -12,7 +12,7 @@ interface IOrderModule is IBasePerpMarket {
         /// Size to modify when settled
         int128 sizeDelta;
         /// block.timestamp of when the order was committed
-        uint256 commitmentTime;
+        uint64 commitmentTime;
         /// The max acceptable price tolerance for settlement
         uint256 limitPrice;
         /// A tip in USD to pay for settlement keepers
@@ -37,7 +37,7 @@ interface IOrderModule is IBasePerpMarket {
     event OrderCommitted(
         uint128 indexed accountId,
         uint128 indexed marketId,
-        uint256 commitmentTime,
+        uint64 commitmentTime,
         int128 sizeDelta,
         uint256 estimatedOrderFee,
         uint256 estimatedKeeperFee
